requirements.txt
Flask==3.0.3
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2
python-dotenv==1.0.1
openai==1.51.2